requirements.txt
# 基础依赖
requests>=2.31.0
Pillow>=10.0.0

# 可选：如果你之后需要打包成 exe 或进行更复杂的异步处理
# pandas>=2.0.0